move static backend functions to BackendCache so there is only one copy
