+++ src/ngraph/runtime/backend.hpp
@@ -81,11 +81,6 @@ namespace ngraph
             void validate_call(std::shared_ptr<const Function> func,
                                const std::vector<std::shared_ptr<runtime::TensorView>>& outputs,
                                const std::vector<std::shared_ptr<runtime::TensorView>>& inputs);
-
-        private:
-            static void* open_shared_library(std::string type);
-            static std::unordered_map<std::string, std::shared_ptr<Backend>>& get_backend_map();
-            static std::unordered_map<std::string, void*> s_open_backends;
         };
     }
 }
